refactor: move vm from core to cli package
