+++ packages/cli/test/tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "extends": "../tsconfig.json",
+    "compilerOptions": {
+        "rootDir": ".."
+    },
+    "include": ["*.ts", "**/*.ts", "../src/**/*.ts"]
+}
