+++ packages/core/package.json
@@ -49,7 +49,6 @@
         "rxjs": "5.5.11",
         "uritemplate": "0.3.4",
         "uuid": "^7.0.3",
-        "vm2": "^3.9.6",
         "web-streams-polyfill": "^3.0.1"
     },
     "scripts": {
